refactor: extract ALE validation gate CLI command
