+++ loopx/cli_commands/__init__.py
@@ -36,6 +36,10 @@ from .agents_last_exam_task_material import (
     handle_agents_last_exam_task_material_command,
     register_agents_last_exam_task_material_commands,
 )
+from .agents_last_exam_validation_gate import (
+    handle_agents_last_exam_validation_gate_command,
+    register_agents_last_exam_validation_gate_commands,
+)
 from .agentissue_runner_flow import (
     handle_agentissue_runner_flow_command,
     register_agentissue_runner_flow_commands,
@@ -145,6 +149,7 @@ __all__ = [
     "handle_agents_last_exam_local_plan_command",
     "handle_agents_last_exam_runner_source_command",
     "handle_agents_last_exam_task_material_command",
+    "handle_agents_last_exam_validation_gate_command",
     "handle_agentissue_runner_flow_command",
     "handle_benchmark_boundary_command",
     "handle_benchmark_command",
@@ -198,6 +203,7 @@ __all__ = [
     "register_agents_last_exam_local_plan_commands",
     "register_agents_last_exam_runner_source_commands",
     "register_agents_last_exam_task_material_commands",
+    "register_agents_last_exam_validation_gate_commands",
     "register_agentissue_runner_flow_commands",
     "register_benchmark_boundary_commands",
     "register_benchmark_command_group",
